requirements.txt
# Core FastAPI
fastapi==0.116.1
uvicorn==0.35.0

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# Data validation
pydantic==2.11.7
pydantic_core==2.33.2

# HTTP & Forms
python-multipart==0.0.20
starlette==0.47.3

# Environment & Configuration
python-dotenv==1.1.1

# Cryptography
cryptography==45.0.7
cffi==1.17.1

# Utilities
click==8.2.1
colorama==0.4.6
h11==0.16.0
idna==3.10
sniffio==1.3.1
anyio==4.10.0

# Type annotations
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.1

# Security libraries
ecdsa==0.19.1
pyasn1==0.6.1
pycparser==2.22
rsa==4.9.1
six==1.17.0